server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.18.4",
    "boto3>=1.42.56",
    "chromadb>=1.5.1",
    "fastapi>=0.133.0",
    "langchain>=1.2.10",
    "langchain-google-genai>=4.2.1",
    "langchain-openai>=1.1.10",
    "pandas>=3.0.1",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "python-magic>=0.4.27",
    "sqlalchemy>=2.0.47",
    "uvicorn>=0.41.0",
]
